Refactor: 공용 UI 컴포넌트 import를 @/components/ui barrel로 통일
- 17개 콜러 파일에서 deep import(@/components/ui/X/X)를 barrel(@/components/ui)로 일괄 정리
- 기능 변화 없음 — barrel은 모든 컴포넌트를 named export로 제공하므로 default(Modal/BottomSheet/Pagination)도 named import 형태로 통일
- 새 코드 작성 시 ui-components 스킬 자동 로딩과 결합해 디자인 시스템 진입점 일관성 확보
