+++ src/app/(content)/news/notices/_component/NoticeDrawer.tsx
@@ -6,8 +6,7 @@ import clsx from 'clsx';
 import { IoCloseOutline, IoChevronBack, IoChevronForward, IoEyeOutline } from 'react-icons/io5';
 import { BsPaperclip } from 'react-icons/bs';
 import useScrollLock from '@/hooks/useScrollLock';
-import { ListItem } from '@/components/ui/ListItem/ListItem';
-import { Label } from '@/components/ui/Label/Label';
+import { ListItem, Label } from '@/components/ui';
 import { NOTICE_CATEGORIES } from '@/constants/notice';
 import { formattedDate, isRecent } from '@/utils/date';
 import type { NoticeType } from '@/types/notice';
