+++ src/app/(content)/sermons/_component/SermonListPage/SermonFilteredList.tsx
@@ -1,5 +1,5 @@
 import GridCard from '../GridCard/GridCard';
-import { EmptyState } from '@/components/ui/EmptyState/EmptyState';
+import { EmptyState } from '@/components/ui';
 import type { SermonWithRelations } from '@/types/sermon';
 import styles from './SermonListPage.module.scss';
